+++ integration_tests/test_replay_block.py
@@ -2,7 +2,6 @@ from pathlib import Path
 
 import pytest
 import web3
-from pystarport import ports
 from web3._utils.method_formatters import receipt_formatter
 from web3.datastructures import AttributeDict
 
@@ -14,7 +13,6 @@ from .utils import (
     deploy_contract,
     sign_transaction,
     supervisorctl,
-    wait_for_port,
 )
 
 
@@ -37,19 +35,23 @@ def test_replay_block(custom_cronos):
     )
     iterations = 400
     gas_limit = 800000
+    gas_price = 100000000000
     for i in range(10):
         nonce = w3.eth.get_transaction_count(ADDRS["validator"])
+        begin_balance = w3.eth.get_balance(ADDRS["validator"])
         txs = [
             contract.functions.test(iterations).buildTransaction(
                 {
                     "nonce": nonce,
                     "gas": gas_limit,
+                    "gasPrice": gas_price,
                 }
             ),
             contract.functions.test(iterations).buildTransaction(
                 {
                     "nonce": nonce + 1,
                     "gas": gas_limit,
+                    "gasPrice": gas_price,
                 }
             ),
         ]
@@ -58,12 +60,8 @@ def test_replay_block(custom_cronos):
             for tx in txs
         ]
         receipt1 = w3.eth.wait_for_transaction_receipt(txhashes[0])
-        try:
-            receipt2 = w3.eth.wait_for_transaction_receipt(txhashes[1], timeout=10)
-        except web3.exceptions.TimeExhausted:
-            # expected exception, tx2 is included but failed.
-            receipt2 = None
-            break
+        # the tx2 should be included in json-rpc response too.
+        receipt2 = w3.eth.wait_for_transaction_receipt(txhashes[1], timeout=10)
         if receipt1.blockNumber == receipt2.blockNumber:
             break
         print(
@@ -73,9 +71,34 @@ def test_replay_block(custom_cronos):
         )
     else:
         assert False, "timeout"
-    assert not receipt2
+
+    # the first tx succeds.
+    assert receipt1.status == 1
+    assert receipt1.gasUsed < gas_limit
+    assert receipt1.cumulativeGasUsed == receipt1.gasUsed
+
+    # the second tx should fail and cost the whole gasLimit
+    assert receipt2.status == 0
+    assert receipt2.gasUsed == gas_limit
+    assert receipt2.cumulativeGasUsed == receipt1.cumulativeGasUsed + gas_limit
+
+    # check get block apis
+    assert w3.eth.get_block(receipt1.blockNumber).transactions == [
+        receipt1.transactionHash,
+        receipt2.transactionHash,
+    ]
+    assert (
+        w3.eth.get_transaction_by_block(receipt1.blockNumber, 1).hash
+        == receipt2.transactionHash
+    )
+
     # check sender's nonce is increased twice, which means both txs are executed.
     assert nonce + 2 == w3.eth.get_transaction_count(ADDRS["validator"])
+    # check sender's balance is deducted as expected
+    assert receipt2.cumulativeGasUsed * gas_price == begin_balance - w3.eth.get_balance(
+        ADDRS["validator"]
+    )
+
     rsp = w3.provider.make_request(
         "cronos_replayBlock", [hex(receipt1.blockNumber), False]
     )
@@ -111,19 +134,10 @@ def test_replay_block(custom_cronos):
     assert replay_receipts[1].status == 0
     assert replay_receipts[1].gasUsed == gas_limit
 
-    # patch the unlucky tx with the new cli command
+    # try to patch the unlucky tx, should have nothing to patch.
     # stop the node0
     end_height = cli.block_height()
     supervisorctl(custom_cronos.base_dir / "../tasks.ini", "stop", "cronos_777-1-node0")
     cli = custom_cronos.cosmos_cli()
     results = cli.fix_unlucky_tx(begin_height, end_height)
-    # the second tx is patched
-    assert results[0][1] == txhashes[1].hex()
-    # start the node0 again
-    supervisorctl(
-        custom_cronos.base_dir / "../tasks.ini", "start", "cronos_777-1-node0"
-    )
-    # wait for tm-rpc port
-    wait_for_port(ports.rpc_port(custom_cronos.base_port(0)))
-    # check the tx indexer
-    assert len(cli.txs(f"ethereum_tx.ethereumTxHash={txhashes[1].hex()}")["txs"]) == 1
+    assert not results